% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_AMI_calculations.R
\name{f_rez}
\alias{f_rez}
\title{Calculate ami and nmi}
\usage{
f_rez(v1, v2)
}
\arguments{
\item{v1}{first cluster vector}

\item{v2}{second cluster vector}
}
\value{
Returns ami and nmi values (a value bewteen 0 and 1) that measures how close the classification between the two clusters is. A value closer to 1 means the clusters are more similar across v1 and v2.
}
\description{
Calculate ami and nmi
}
\examples{
cluster1 <- sample(1:5, size = 10, replace = TRUE)
cluster2 <- sample(1:2, size = 10, replace = TRUE)
f_rez( cluster1, cluster2 )
}
