% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_generate_test_data.R
\name{generateSimulationData}
\alias{generateSimulationData}
\title{generateSimulationData}
\usage{
generateSimulationData(sigma = 1, sparsity = 0.02, seed = 1234)
}
\arguments{
\item{sigma}{noise level - default is 1}

\item{sparsity}{percentage of relevent features - default is 0.02}

\item{seed}{Random seed - default is 1234}
}
\value{
Returns simulation data - a 21000x1000 sparse matrix with 4 clusters (2 clusters are n = 10,000 and 2 clusteraer n = 500)
}
\description{
generateSimulationData
}
\examples{
X = generateSimulationData()
}
