% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_documentation.R
\docType{data}
\name{OV}
\alias{OV}
\title{OV dataset}
\format{
A data frame with 20530 rows and 306 columns (will be transposed in TCGA_cleanData())
}
\source{
https://github.com/compgenome365/TCGA-Assembler-2
}
\usage{
OV
}
\description{
A dataset containing genetic information for patients with ovarian cancer
}
\keyword{datasets}
