% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmImplementation.R
\name{RJclust}
\alias{RJclust}
\title{RJclust}
\usage{
RJclust(X, num_cut, seed = 1)
}
\arguments{
\item{X}{Data input (if TCGA data, data needs to be pre-procesed)}

\item{num_cut}{Number of cuts for RJ algorithm (suggested sqrt(p))}

\item{seed}{Seed (defalt = 1)}
}
\value{
Returns RJ algorithm result
}
\description{
RJclust
}
\examples{
data(OV)
X = TCGA_cleanData(OV)
clust = RJclust(X, 3)
}
