% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapperFunctions.R
\name{scaleRJ}
\alias{scaleRJ}
\title{scaleRJ}
\usage{
scaleRJ(X, medians = FALSE)
}
\arguments{
\item{X}{Numeric matrix to be scaled}

\item{medians}{A logical, if TRUE the medians of the columns are used to scaled}
}
\value{
The scaled matrix
}
\description{
scaleRJ
}
\examples{
data = generateSimulationData()
Xscaled = scaleRJ(data$X)
}
