% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_clean_TCGA.R
\name{TCGA_cleanData}
\alias{TCGA_cleanData}
\title{TCGA_cleanData}
\usage{
TCGA_cleanData(data, threshold = 50)
}
\arguments{
\item{data}{Data to preprocess}

\item{threshold}{Percentage of 0s that will cause a gene (column) to be discarded (default is 50)}
}
\value{
Preprocessed data by transposing matrix and removing genes (columns in X matrix) that have a percentage of gene values as 0 over given threshold
}
\description{
TCGA_cleanData
}
\examples{
data(OV)
data_preprocessed = TCGA_cleanData(OV, 50)

print(dim(OV))
print(dim(data_preprocessed))
}
