% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmImplementation.R
\name{TCGA_RJclust_removePatients}
\alias{TCGA_RJclust_removePatients}
\title{TCGA_RJclust_removePatients}
\usage{
TCGA_RJclust_removePatients(X, num_cut, numRemove = 15, seed = 1)
}
\arguments{
\item{X}{TCGA preprocessed data}

\item{num_cut}{Number of cuts for RJ algorithm (suggested sqrt(p))}

\item{numRemove}{Number of patients to remove}

\item{seed}{Seed (defalt = 1)}
}
\value{
Returns RJ algorithm result
}
\description{
TCGA_RJclust_removePatients
}
\examples{
data(OV)
X = TCGA_cleanData(OV)
clustNew = TCGA_RJclust_removePatients(X, 3)
}
