\name{RJcluster-package}
\alias{RJcluster-package}
\alias{RJcluster}
\docType{package}
\title{
\packageTitle{RJcluster}
}
\description{
\packageDescription{RJcluster}
}
\details{

  \tabular{ll}{
    Package: \tab RJcluster\cr
    Type: \tab Package\cr
    Version: \tab 3.2.4\cr
    Date: \tab 07-15-2021\cr
    License: \tab GPL>=2\cr
  }
}
\author{
\packageAuthor{RJcluster}

Maintainer: \packageMaintainer{RJcluster}
}